+++ game-digimonstorytimestranger/CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## Future Changes (NOT IMPLEMENTED YET)
 
-- None Planned
+- Notification to install .NET 9.0.4 SDK (x64 and x86) + may need to launch as admin once to download dependencies.
 
-## [0.1.0] - 2025-XX-XX
+## [0.1.1] - 2025-10-09
+
+- Added button to open Saves folder ("gamedata/savedata").
+- Improved setup notification text.
+
+## [0.1.0] - 2025-10-04
 
 - Initial Release